requirements.txt
grpcio==1.39.0
grpcio-tools==1.39.0
protobuf==3.17.3
six==1.16.0
elementpath==2.5.0
xmlschema==1.10.0
